#1418 RedundantFieldInitializer false positive with large long value

+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/optimizations/RedundantFieldInitializerRule.java
@@ -100,7 +100,7 @@ public class RedundantFieldInitializerRule extends AbstractJavaRule {
                                     value = (int) literal.getImage().charAt(1);
                                 }
 
-                                if (value.intValue() == 0) {
+                                if (value.longValue() == 0) {
                                     addViolation(data, variableDeclarator);
                                 }
                             }
